added velocity change functionality with correct dynamic reconfigure client name, added python rosservice call of grippe toggle, NEED TO BE REFACTORED INTO STATE MACHINE
